requirements.txt
pandas
Jinja2
python-socks
lxml
tqdm
goslate